+++ in-adapter-api/src/main/java/gohigher/oauth2/user/GoogleOAuth2User.java
@@ -1,4 +1,4 @@
-package gohigher.user.oauth2;
+package gohigher.oauth2.user;
 
 import java.util.Map;
