+++ in-adapter-api/src/main/java/gohigher/oauth2/user/OAuth2UserInfoFactory.java
@@ -1,6 +1,8 @@
-package gohigher.user.oauth2;
+package gohigher.oauth2.user;
 
 import java.util.Map;
+
+import gohigher.user.oauth2.Provider;
 
 public class OAuth2UserInfoFactory {
